requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.3.0
attrs==23.2.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
cpe==1.2.1
cybox==2.1.0.21
deepdiff==6.7.1
drawsvg==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fqdn==1.5.1
idna==3.6
isoduration==20.11.0
jsonpointer==2.4
jsonschema==4.17.3
loguru==0.7.2
lxml==5.0.1
maec==4.1.0.17
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mitreattack-python==3.0.2
mixbox==1.0.5
netaddr==0.10.1
numpy==1.26.3
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
pluralizer==1.2.0
pooch==1.8.0
pycountry==23.12.11
Pygments==2.17.2
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
simplejson==3.19.2
six==1.16.0
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.1.4
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
tqdm==4.66.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.1.0
weakrefmethod==1.0.3
webcolors==1.13
XlsxWriter==3.1.9
